Add support for non-basic rule functions.
Non-basic rule functions are functions that, besides returning
the boolean result of the validation, also return an error value.

Non-basic rule functions are inteded to be used for cases where
the validation can potentially fail due to other reasons than just
"invalid input". For example a validation function that has a db
dependency but the db connection fails, in a case like this the
non-basic rule function can return the connection error.

When a non-basic rule function returns a non-`nil` error value,
then that error value will be returned immediately regardless of
whether or not an error aggregator is being used.